Shifted explicit conversions using strtod and atoi to use templated as calls on the yaml node object

+++ src/sleep.hpp
@@ -23,6 +23,6 @@ namespace mwg {
         
     private:
         // possible next states with probabilities
-        long long sleeptime;
+        uint64_t sleeptime;
 };
 }
